+++ dev/index.ts
@@ -1,4 +1,5 @@
 import fs from "node:fs";
+import { BootstrapTheme } from "../src";
 
 const Bootstrap5Generator = require("../").Bootstrap5Generator
 
@@ -26,14 +27,26 @@ const variables = {
 	"$color-2": "lighten($color-1,20% )"
 }
 
-const theme = {
-	variables,
+const theme: BootstrapTheme = {
+	variables: {
+
+		"$body-bg": "#ffff00",
+		"$info": "#99b6c3",
+		"$badge-pill-border-radius": "20rem",
+		"$border-radius": "1.5rem",
+		"$border-radius-lg": "1.5rem",
+		"$border-radius-sm": "1.5rem",
+		"$toast-border-radius": "2rem"
+	},
+	components: [
+
+	],
 	colors: {
 		primary: "#FFaaCC",
 		secondary: "#ccdbe1",
 		warning: "#ee8140",
 		danger: "#CC3105",
-		//	success: "#B6E186",
+		success: "#B6E186",
 		highlight: "#FF00FF"
 	}
 }