+++ CHANGELOG.md
@@ -5,9 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Added support to choose which components to import when generating Bootstrap css
 
 ## [0.4.2] - 2024-04-17
 - Added missing import when generating CSS
+
 ## [0.4.1] - 2024-04-17
 - Fixed bug on method getHashFromTheme